Add bypass_auth_check for local dev
