+++ .env.example
@@ -1,2 +1,3 @@
 ALLOWED_ORIGINS=[]
 API_KEY=
+BYPASS_AUTH_CHECK=true
